Refactor: move card validation utils
